requirements.txt
gradio
openai-whisper
spaces
torch